[step4] refactor: 패키지 변경, 관련 코드 수정
