+++ src/main/kotlin/racingcar/utils/InputParser.kt
@@ -1,4 +1,4 @@
-package racingcar
+package racingcar.utils
 
 object InputParser {
