+++ src/test/kotlin/racingcar/NumberInputTest.kt
@@ -6,6 +6,7 @@ import org.junit.jupiter.api.Test
 import org.junit.jupiter.params.ParameterizedTest
 import org.junit.jupiter.params.provider.NullSource
 import org.junit.jupiter.params.provider.ValueSource
+import racingcar.utils.InputParser
 
 class NumberInputTest {
     @Test
